requirements.txt
textblob
spacy
transformers
torch
scikit-learn
nltk
snscrape
instaloader
praw
tweepy
facebook-scraper
networkx
pandas
numpy
matplotlib
seaborn
plotly
selenium
beautifulsoup4
requests
urllib3
streamlit
streamlit-option-menu
python-dotenv
python-dateutil
tqdm
joblib